SecurityTest/webapp/requirements.txt
aniso8601==9.0.1
autopep8==2.0.2
click==8.1.3
colorama==0.4.6
flake8==6.0.0
Flask==2.2.3
Flask-RESTful==0.3.9
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mccabe==0.7.0
mysql-connector-python==8.0.32
packaging==23.0
platformdirs==3.2.0
protobuf==3.20.3
pycodestyle==2.10.0
pyflakes==3.0.1
pytoolconfig==1.2.5
pytz==2023.3
rope==1.7.0
six==1.16.0
tomli==2.0.1
Werkzeug==2.2.3
